pyproject.toml
[tool.poetry]
name = "bewise"
version = "0.1.0"
description = ""
authors = ["d.shinkarev <d.shinkarev@vkteam.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.6"
psycopg2-binary = "^2.9.10"
kafka-python = "^2.0.2"
python-dotenv = "^1.0.1"
uvicorn = {extras = ["standart"], version = "^0.34.0"}
sqlalchemy = "^2.0.37"
alembic = "^1.14.0"
pydantic-settings = "^2.7.1"
aiokafka = "^0.12.0"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
